+++ src/Main/AW.h
@@ -3,7 +3,7 @@
 // * Description:    Audio Wizard Header File                                * //
 // * Author:         TT                                                      * //
 // * Website:        https://github.com/The-Wizardium/Audio-Wizard           * //
-// * Version:        0.5.0                                                   * //
+// * Version:        0.6.0                                                   * //
 // * Dev. started:   12-12-2024                                              * //
 // * Last change:    01-09-2025                                              * //
 /////////////////////////////////////////////////////////////////////////////////
@@ -15,6 +15,14 @@
 #include "AW_Main.h"
 #include "AW_Peakmeter.h"
 #include "AW_Waveform.h"
+
+
+////////////////////////////////////////
+// * AUDIO WIZARD COMPONENT VERSION * //
+////////////////////////////////////////
+#pragma region Component Version
+inline constexpr const char* AW_COMPONENT_VERSION = "0.6.0";
+#pragma endregion
 
 
 //////////////////////