+++ src/Main/AW_Settings.h
@@ -3,7 +3,7 @@
 // * Description:    Audio Wizard Settings Header File                       * //
 // * Author:         TT                                                      * //
 // * Website:        https://github.com/The-Wizardium/Audio-Wizard           * //
-// * Version:        0.5.0                                                   * //
+// * Version:        0.6.0                                                   * //
 // * Dev. started:   12-12-2024                                              * //
 // * Last change:    23-12-2025                                              * //
 /////////////////////////////////////////////////////////////////////////////////
@@ -158,7 +158,7 @@ public:
 	static constexpr int monitorDisplayMetricsModeDefault = 4;
 	static constexpr int monitorDisplayRefreshRateDefault = 33;
 
-	static constexpr bool systemDebugLogDefault = false;
+	static constexpr bool systemDebugLogDefault = true;
 
 	// * STATIC MEMBERS * //
 	static inline cfg_bool analysisDisplayColors{ guid_analysisDisplayColors, analysisDisplayColorsDefault };